Move some files to new packages in the model server
