+++ model-server/src/main/kotlin/org/modelix/model/server/handlers/HistoryHandler.kt
@@ -1,8 +1,7 @@
-package org.modelix.model.server
+package org.modelix.model.server.handlers
 
 import io.ktor.http.*
 import io.ktor.server.application.*
-import io.ktor.server.auth.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
