+++ model-server/src/main/kotlin/org/modelix/model/server/store/LocalModelClient.kt
@@ -12,7 +12,7 @@
  * specific language governing permissions and limitations
  * under the License. 
  */
-package org.modelix.model.server
+package org.modelix.model.server.store
 
 import org.apache.commons.collections4.IterableUtils
 import org.modelix.model.IKeyListener
