+++ model-server/src/test/kotlin/org/modelix/model/server/JsonAPITest.kt
@@ -23,7 +23,12 @@ import io.ktor.server.websocket.*
 import org.json.JSONObject
 import org.modelix.authorization.installAuthentication
 import org.modelix.model.api.ITree
-import org.modelix.model.server.*
+import org.modelix.model.server.handlers.JsonModelServer
+import org.modelix.model.server.handlers.asObjectList
+import org.modelix.model.server.handlers.buildJSONArray
+import org.modelix.model.server.handlers.buildJSONObject
+import org.modelix.model.server.store.InMemoryStoreClient
+import org.modelix.model.server.store.LocalModelClient
 import kotlin.test.*
 
 class JsonAPITest {
